fix(*): update the test cases
